Split parsers and models by package. Update test cases md. Remove extra print statements
