+++ watchdogs/io/parsers/EncodingViewerArgs.py
@@ -5,7 +5,7 @@
 from collections import OrderedDict
 
 from watchdogs.base.models.Common import Common
-from watchdogs.io.models.FileArgs import FileArgs
+from watchdogs.io.parsers import FileArgs
 from watchdogs.utils.Constants import (EMPTY, LFN)
