+++ watchdogs/web/models/__init__.py
@@ -1,1 +1,2 @@
 from .AVI import AVI
+from FileFuzzer import FileFuzzer
