+++ watchdogs/web/parsers/FileFuzzerArgs.py
@@ -4,7 +4,7 @@
 
 import argparse
 
-from watchdogs.io.models import FileArgs
+from watchdogs.io.parsers import FileArgs
 from watchdogs.base.models import Common
 from watchdogs.utils.Constants import (EMPTY, COLON)
