+++ watchdogs/web/parsers/__init__.py
@@ -1,0 +1,1 @@
+from .FileFuzzerArgs import FileFuzzerArgs
